% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme_points.R
\name{extreme_points}
\alias{extreme_points}
\title{Extreme Points and Zero Crossings}
\usage{
extreme_points(y, ndata = length(y))
}
\arguments{
\item{y}{Signal values (Numeric | vector)}

\item{ndata}{Number of data points (Integer)}
}
\value{
Index values for minima, maxima, zero crossings, number of extremas and zero crossing
}
\description{
Finding Local Extrema and Zero-crossings
}
\details{
A local extrema is the point at which a maximum or minimum value of the function in some open interval containing the point is obtained. A zero-crossing is a point where the sign of a mathematical function changes.
}
\examples{
x=1:100
y = 2 / 3 * sin(x * 30) + 2 / 3 * sin(x * 17.5) + 4 / 5 *cos(x * 2)
extreme_points(y)
}
\references{
\url{https://cran.r-project.org/web/packages/EMD/index.html}
}
\keyword{extrema}
\keyword{local}
\keyword{maxima}
\keyword{minima}
