% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_extrema.R
\name{find_extrema}
\alias{find_extrema}
\title{Find Extreme Points}
\usage{
find_extrema(signal, INCLUDE_ENDPOINTS = TRUE)
}
\arguments{
\item{signal}{Signal values (Numeric | vector)}

\item{INCLUDE_ENDPOINTS}{whether to include end points or not (Boolean)}
}
\value{
Index of all extrema values (including starting and ending points)
}
\description{
Method for finding Extreme Points
}
\details{
A local extrema is the point at which a maximum or minimum value of the function in some open interval containing the point is obtained.
}
\examples{
signal=c( 0.841471 ,0.9092974,0.14112,-0.7568025,-0.9589243)
find_extrema(signal)
}
\author{
Shubhra Prakash, \email{shubhraprakash247@gmail.com}
}
\keyword{extrema}
