% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmd.R
\name{lmd}
\alias{lmd}
\title{Local Mean Decomposition}
\usage{
lmd(
  signal,
  INCLUDE_ENDPOINTS = TRUE,
  MAX_SMOOTH_ITERATION = 12,
  MAX_ENVELOPE_ITERATION = 200,
  ENVELOPE_EPSILON = 0.01,
  CONVERGENCE_EPSILON = 0.01,
  MAX_NUM_PF = 8
)
}
\arguments{
\item{signal}{Signal values (Numeric | vector)}

\item{INCLUDE_ENDPOINTS}{Whether to treat the endpoint of the signal as a pseudo-extreme point (Boolean)}

\item{MAX_SMOOTH_ITERATION}{Maximum number of iterations of moving average algorithm (Integer)}

\item{MAX_ENVELOPE_ITERATION}{Maximum number of iterations when separating local envelope signals (Integer)}

\item{ENVELOPE_EPSILON}{Terminate processing when obtaining pure FM signal (Double)}

\item{CONVERGENCE_EPSILON}{Terminate processing when modulation signal converges (Double)}

\item{MAX_NUM_PF}{The maximum number of PFs generated(Integer)}
}
\value{
list(pf,residue) | PFs:The decompose functions arranged from high frequency to low frequency | residue:residual component
}
\description{
Method for finding Product Functions (PFs)
}
\details{
LMD is a method of decomposing signal into Product Functions (PFs) based on algorithm presented in Jonathan S. Smith. The local mean decomposition and its application to EEG perception data. Journal of the Royal Society Interface, 2005, 2(5):443-454
}
\examples{
x=1:100
y = 2 / 3 * sin(x * 30) + 2 / 3 * sin(x * 17.5) + 4 / 5 *cos(x * 2)
plot(y,type="l")
lmd(y)
}
\references{
\url{https://pypi.org/project/PyLMD/}
}
\author{
Shubhra Prakash, \email{shubhraprakash247@gmail.com}
}
\keyword{lmd}
