% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmd_shiny.R
\name{lmd_shiny}
\alias{lmd_shiny}
\title{LMD Shiny}
\usage{
lmd_shiny()
}
\description{
Shiny Dashboard for plotting Product Functions (PFs) and Residue
}
\author{
Shubhra Prakash, \email{shubhraprakash279@gmail.com}
}
\keyword{Shiny}
