% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_mean_and_envelope.R
\name{local_mean_and_envelope}
\alias{local_mean_and_envelope}
\title{Local Mean and Envelope}
\usage{
local_mean_and_envelope(signal, extrema)
}
\arguments{
\item{signal}{Signal values (Numeric | vector)}

\item{extrema}{indexes for extreme values}
}
\value{
mean, envelope and smoothed mean and envelope values
}
\description{
Method for finding Local Mean and Envelope
}
\examples{
signal = sin(1:10)
extrema = c(1 , 2,  5,  8, 10)
local_mean_and_envelope(signal, extrema)
}
\references{
\url{https://pypi.org/project/PyLMD/}
}
\author{
Shubhra Prakash, \email{shubhraprakash279@gmail.com}
}
\keyword{Envelope}
\keyword{Local}
\keyword{Mean}
\keyword{and}
