% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_average_smooth.R
\name{moving_average_smooth}
\alias{moving_average_smooth}
\title{Weighted Moving Average}
\usage{
moving_average_smooth(signal, window, MAX_SMOOTH_ITERATION = 12)
}
\arguments{
\item{signal}{Signal values (Numeric | vector)}

\item{window}{filter weights for smoothing (Numeric | vector)}

\item{MAX_SMOOTH_ITERATION}{Maximum number of iterations of moving average algorithm (Integer)}
}
\value{
smooth signal
}
\description{
Weighted Moving Average Smoothing
}
\details{
Weighted Moving Average Smoothing is used to smoothen the mean and envelope signal
}
\examples{
x=0:100
y = 2 / 3 * sin(x * 30) + 2 / 3 * sin(x * 17.5) + 4 / 5 *cos(x * 2)
plot(y,type="l")
wma=moving_average_smooth(y,5)
}
\references{
\url{https://pypi.org/project/PyLMD/}
}
\author{
Shubhra Prakash, \email{shubhraprakash247@gmail.com}
}
\keyword{Average}
\keyword{Moving}
\keyword{Weighted}
\keyword{wma}
