% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square_val.R
\name{square_val}
\alias{square_val}
\title{Square Val}
\usage{
square_val(x = 2)
}
\arguments{
\item{x}{# An integer}
}
\value{
y: An Integer
}
\description{
Square Val
}
\examples{
square_val(2)
}
\references{
\url{https://en.wikipedia.org/wiki/List_of_Crayola_crayon_colors}
}
\seealso{
\code{\link{utils}}
}
\author{
Shubhra Prakash, \email{shubhraprakash247@gmail.com}
}
\keyword{hplot}
